Add containers to change pages
