+++ frontend/package.json
@@ -7,9 +7,14 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "react": "^16.13.1",
+    "react-calendar": "^3.1.0",
     "react-dom": "^16.13.1",
+    "react-draggable": "^4.4.3",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
-    "react-swipeable-views": "^0.13.9"
+    "react-swipeable": "^5.5.1",
+    "react-swipeable-views": "^0.13.9",
+    "react-trello": "^2.2.6"
   },
   "scripts": {
     "start": "react-scripts start",
