+++ frontend/src/containers/Project.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Project = () => (
+  <div>
+    <h2>Project Page</h2>
+  </div>
+);
+
+export default Project;
